pyproject.toml
[project]
name = "libreoffice-env"
version = "0.1.0"
description = "MCP-based sandboxed RL environment for LibreOffice tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.22.0",
    "docker>=7.0.0",
    "odfpy>=1.4.1",
    "pandas>=2.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "sympy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
